pyproject.toml
[tool.poetry]
name = "aws-alarm-notifier"
version = "0.1.0"
description = "Sends notifications of AWS CloudFront alarm state changes to TBG Slack channels"
authors = ["Justin Hendrickson <justin.hendrickson@thebernardgroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.104.0"
tbg-cdk-python = "^7.0.5"
tbg-cdk-nag = "^7.0.2"


[[tool.poetry.source]]
name = "tbg"
url = "https://tbg-538493872512.d.codeartifact.us-east-1.amazonaws.com/pypi/python/simple/"
priority = "default"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"


[tool.poetry.group.handler.dependencies]
aws-lambda-powertools = {extras = ["parser", "tracer"], version = "^2.26.0"}
python-json-logger = "^2.0.7"
pynamodb = "^5.5.0"
requests = "^2.31.0"
sentry-sdk = "^1.34.0"
slack-sdk = "^3.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
